// frontend/src/core/api/generated/model/rosterEntryRequest.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Constructor ERP API
 * OpenAPI spec version: 1.0
 */
import type { RosterEntryRequestKind } from './rosterEntryRequestKind'

export interface RosterEntryRequest {
  kind: RosterEntryRequestKind
  /**
   * @minLength 0
   * @maxLength 120
   */
  name: string
  /**
   * @minLength 0
   * @maxLength 60
   * @nullable
   */
  trade?: string | null
  /**
   * @minLength 0
   * @maxLength 20
   * @nullable
   */
  phone?: string | null
}